Hopefully this wil commit properly and fix your issue

+++ src/Mod/src/main/java/net/mcreator/mindustryinminecraft/procedures/ConveyermakerkeyOnKeyPressedProcedure.java
@@ -1,22 +1,8 @@
 package net.mcreator.mindustryinminecraft.procedures;
 
-import net.minecraftforge.fmllegacy.network.NetworkHooks;
+import net.minecraftforge.eventbus.api.Event;
 
-import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.MenuProvider;
-import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.network.chat.TextComponent;
-import net.minecraft.network.chat.Component;
-import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.core.BlockPos;
-
-import net.mcreator.mindustryinminecraft.world.inventory.CraftingScreenMenu;
-
-import io.netty.buffer.Unpooled;
+import javax.annotation.Nullable;
 
 public class ConveyermakerkeyOnKeyPressedProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z, Entity entity) {
@@ -24,7 +10,7 @@ public class ConveyermakerkeyOnKeyPressedProcedure {
 			return;
 		{
 			if (entity instanceof ServerPlayer _ent) {
-				BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bpos = new BlockPos(x, y, z);
 				NetworkHooks.openGui((ServerPlayer) _ent, new MenuProvider() {
 					@Override
 					public Component getDisplayName() {
